requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1
requests==2.32.5

# Utils
django-debug-toolbar==4.2.0
Pillow==11.1.0
django-storages==1.14.6
boto3==1.34.162
zipstream-ng==1.7.1

# Services
celery==5.3.4
redis==5.0.1
hiredis==3.1.0

# Tooling
django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
drf-spectacular==0.26.5

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0

astor==0.8.1 
django-jazzmin==3.0.1

# AI
anthropic==0.34.2

# DB Layer
psycopg[binary]==3.2.13
# mysqlclient==2.2.0
